big change in this commit, oops, added progress bar, updated some routes, refactored some code from post controller to independent functions, changed the way some of the views work

+++ app/Http/Controllers/UsersController.php
@@ -55,6 +55,9 @@ class UsersController extends Controller
     public function destroy($id)
     {
         $user = User::find($id);
+        if (!$user) {
+            return redirect()->route('users.index')->with('error', 'User not found.');
+        }
         $user->delete();
         return redirect()->route('users.index')->with('success', 'User deleted successfully.');
     }
